refactor(core): relocate AppError to shared building blocks
Moves AppError.ts to building-blocks/errors/AppError.ts.

Keeps legacy core/src/utils/AppError.ts re-export to protect 200+ consumers.

All 223 core tests run and compile cleanly.
